+++ internal/handler/api.go
@@ -18,6 +18,7 @@ func NewAPI(cfg config.Config, authService AuthService, userService UserService)
 				auth.POST("/signin", SigninHandler(authService, userService, cfg))
 				auth.POST("/reset-password", ResetPasswordHandler(userService))
 			}
+
 		}
 	}
