+++ internal/handler/auth.go
@@ -13,7 +13,6 @@ import (
 	"github.com/dilyara4949/flight-booking-api/internal/handler/response"
 	"github.com/dilyara4949/flight-booking-api/internal/middleware"
 	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
 )
 
 type AuthService interface {
@@ -22,7 +21,7 @@ type AuthService interface {
 
 type UserService interface {
 	CreateUser(ctx context.Context, signup request.Signup, password string) (domain.User, error)
-	ResetPassword(ctx context.Context, userID uuid.UUID, newPassword string, requirePasswordReset bool) error
+	ResetPassword(ctx context.Context, req request.ResetPassword, requirePasswordReset bool) error
 	ValidateUser(ctx context.Context, signin request.Signin) (domain.User, error)
 }
 
@@ -95,6 +94,12 @@ func SigninHandler(authService AuthService, userService UserService, cfg config.
 			return
 		}
 
+		if user.RequirePasswordReset {
+			c.JSON(http.StatusForbidden, response.Error{Error: "access denied: reset password required"})
+
+			return
+		}
+
 		token, err := authService.CreateAccessToken(c, user, cfg.JWTTokenSecret, cfg.AccessTokenExpire)
 		if err != nil {
 			slog.Error("signin: error at creating access token,", "error", err.Error())
@@ -128,6 +133,31 @@ func AccessCheck(req gin.Context, expectedContextID, expectedIDKey string) bool 
 	}
 
 	return false
+}
+
+func ResetPasswordHandler(userService UserService) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		var req request.ResetPassword
+
+		err := c.ShouldBind(&req)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, response.Error{Error: "error at binding request body"})
+			return
+		}
+
+		if req.NewPassword == "" || req.OldPassword == "" || req.Email == "" {
+			c.JSON(http.StatusBadRequest, response.Error{Error: response.ErrEmptyRequestFields.Error()})
+			return
+		}
+
+		err = userService.ResetPassword(c, req, false)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, response.Error{Error: err.Error()})
+			return
+		}
+
+		c.JSON(http.StatusOK, "password reset successful")
+	}
 }
 
 func domainUserToResponse(user domain.User) response.User {
