+++ internal/handler/request/signup.go
@@ -6,5 +6,7 @@ type Signup struct {
 }
 
 type ResetPassword struct {
+	Email       string `json:"email"`
+	OldPassword string `json:"old_password"`
 	NewPassword string `json:"new_password"`
 }
